requirements.txt
langgraph
langchain
langchain-openai
langchain-community
tavily-python
chromadb
python-dotenv
langchain-huggingface
sentence-transformers
faiss-cpu
pandas
openpyxl
